CDS: Retry remote front-end transactions on AskTimeoutException
With the front-end PrimaryShardInfo cache, if the cached primary/leader
shard is remote and unavailable, the RemoteTransactionContextSupport will
fail with an AskTimeoutException when it tries to send the
CreateTransaction message. Since it can take at least 1 election timeout
period to re-elect a new leader, I changed
RemoteTransactionContextSupport to also retry on AskTimeoutException (it
already retries on NoShardLeaderException). However instead of
re-sending the CreateTransaction message, as it did before, it now
re-sends the FindPrimary message to get a new primary shard actor.

I also modified how RemoteTransactionContextSupport retries. It will now
retry for a total period of 2 times the shard election timeout which should
be  ample time for a re-election to occur. If no leader is found then the
txn will fail.

For the unit tests, I needed to be able to set the operationTimeout in
millis so that required changes in the DatastoreContext and other
classes.

Change-Id: I3efd3f841cf41b7738aedb694fa18b44851b3074
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>