CDS: Real snapshot log trimming changes
Bug 2692 changed real snapshotting to trim the in-memory log based on
the current replicatedToAllIndex for the normal case when all follower's
are up and "fake" snapshotting is advancing replicatedToAllIndex.

When a follower is down, real snapshots don't trim the log b/c
replicatedToAllIndex isn't advancing unless the memory threshold is
exceeded. So it will let the in-memory log keep growing past the
snapshot batch count. This sort of defeats the purpose of snapshotting,
ie to keep the journal size in check both on disk and in memory. It's
also a bit dangerous - it can chew up a lot of memory and starve the
rest of the system and cause large STW GC's once the follower comes
back up and the log is cleared. This can also cause multiple snapshots
in the follower once it comes back and catches up - eg, if it's behind
60K entries, it will snapshot after each 20K batch in quick succession.

To alleviate the potential excessive memory growth, in addition to
trimming the log from the captured lastAppliedIndex if the log memory
size threshold is exceeded, I changed the code to do the same if the log
size exceeds the snapshot batch count. So if a follower is down long
enough to exceed the snapshot batch count, the leader will install a
single snapshot to catch up the follower. Otherwise, the follower will
be caught up via AppendEntries.

I also fixed an NPE in SnapshotSupportImpl#persist where the
currentCaptureSnapshot was null. Here was the sequence:

2015-07-02 03:49:08,671 | INFO  | Initiating Snapshot Capture..
2015-07-02 03:49:10,287 | INFO  | Persisting of snapshot done...
2015-07-02 03:49:10,620 | INFO  | Initiating Snapshot Capture..
2015-07-02 03:49:10,638 | INFO  | SaveSnapshotSuccess received for snapshot
2015-07-02 03:49:11,809 | INFO  | CaptureSnapshotReply received...
2015-07-02 03:49:11,817 | WARN  | lt-dispatcher-14 | ShardManager |
Supervisor Strategy of resume applied
        at
org.opendaylight.controller.cluster.raft.SnapshotSupportImpl.persist(SnapshotSupportImpl.java:128)
        at
org.opendaylight.controller.cluster.raft.RaftActor.handleCommand(RaftActor.java:370)

This occurred when 2 snapshots were initiated in quick succession. As
soon as the first one persisted, the snapshotCaptureInitiated flag was
cleared allowing the second one to initiate. However, when the first one
saved successfully and was commited, it cleared the cached
currentCaptureSnapshot. When the second capture completed and tried to
persist, it encountered a null currentCaptureSnapshot. So I changed it
to clear snapshotCaptureInitiated in commit instead of persist.

Change-Id: I12284b37c34c970adb05807c0f699e97409ec49b
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>