Update patch set 1
Patch Set 1: Code-Review+1

Since we're moving netconf out of the controller project the aaa features are going to go with it. In the meantime i think it's better to enable it to not have netconf broken on master.

Patch-set: 1
Reviewer: Gerrit User 3127 <3127@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1