BUG 2185 : Make the Custom Raft Behavior externally configurable
A class which implements CustomizableRaftBehavior needs to be set
to customize the behavior. I am hoping that this will be a barrier
to people unintentionally breaking the default raft behavior as
they would need to know the internals of clustering in order to
supply a valid class.

I added two configurable raft behavior classes for the two known
use cases,

- org.opendaylight.controller.cluster.datastore.utils.TestOnlyRaftBehavior
- org.opendaylight.controller.cluster.datastore.utils.TwoNodeClusterRaftBehavior

Change-Id: Ic3cc2f27754c37e85c3be8a863764fc88ec84399
Signed-off-by: Moiz Raja <moraja@cisco.com>