Update patch set 1
Patch Set 1:

you mean use class of event for filtering as opposed to value in event? As I recall, again it has been a while, the important filtering need was around value in an event, not just type of event. Having the granularity even down to "did this field value change" would be better than "did this object change", but again "did this value change to/from X" would be better.

Patch-set: 1