Update patch set 2
Patch Set 2:

(1 comment)

I think there's more work to do here. The Shard's DataChangeListenerSupport delays registrations until it becomes the leader. We don't want that for listeners that want to be notified on each instance. In that case we need to delay it until it becomes a leader or follower. Also for the latter we need to ensure it has a leader.

Patch-set: 2