Update patch set 2
Patch Set 2:

Sorry I missed that originally. That will also register if a Candidate and the listener will be notified of pre-existing data from persistence. I'm not sure we want to notify w/o a leader b/c DCL's typically create transactions which will either be delayed until a leader is elected or fail. So do we want to notify a DCL and possibly set it up for subsequent failure?

Patch-set: 2