Update patch set 3
Patch Set 3:

(9 comments)

We should eliminate the duplication in DefaultAkkaConfigurationReader and FileConfigurationReader. They both do the same thing. I think we should eliminate FileConfigurationReader and ConfigurationReader and change the CDS code to use *AkkaConfigurationReader in sal-clustering-commons. Also move ResourceConfigurationReader to sal-clustering-commons and rename to ResourceAkkaConfigurationReader. And rename DefaultAkkaConfigurationReader to FileAkkaConfigurationReader to make it explicit.

Patch-set: 3