Update patch set 2
Patch Set 2:

Looking in the build log:

https://jenkins.opendaylight.org/releng/job/controller-verify-stable-lithium/jdk=openjdk8,nodes=dynamic_verify/lastFailedBuild/console

It appears that the build failed due to a timeout:

Build timed out (after 360 minutes). Marking the build as failed.
[ssh-agent] Stopped.
Build was aborted
[FINDBUGS] Skipping publisher since build result is FAILURE

I do mean a separate profile in the POM. Try pushing a new patch set with that change, it might result in a successful build next time.

Patch-set: 2
Label: Verified=0