Restconf returns 503 if TokenAuthFilter fails to load before import
org.opendaylight.aaa.sts.TokenAuthFilter is utilized as a restconf
JAX-RS filter.  If TokenAuthFilter class is not loaded prior to
starting the sal_rest_connector, there is a chance that the Class
is not found, resulting in a loading race condition.  This patch
includes code to specifically import the AAA sts package prior to
creating the sal_rest_connector service.

Change-Id: I870fe683be49c529258c9ed98cc8d6f983721ec6
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>