Update patch set 24
Patch Set 24:

Actually, the tests are run in the daily job for ovsdb:

https://jenkins.opendaylight.org/releng/view/ovsdb/job/ovsdb-daily-beryllium/

where they are passing, as well as passing locally against master for me.

To run them, fire up an instance of ovs listening for connections and use something like:

mvn verify -Pintegrationtest -Dovsdbserver.ipaddress=127.0.0.1 -Dovsdbseer.port=16640

The error I was getting was that the variable 'db' (which is always the databroker) is null... this value is injected in via the config subsystem, thus my hypothesis that something is broken causing it to be injected with an incorrect null value.

Patch-set: 24