Bug-4214 - Add support for configurable snapshot chunk size.
Porting bug-fix to master branch manually as the cherry-pick is failing.

Added a new variable in the distributed-datastore-provider.yang. This
will be used to configure the snapshot chunk size. Added various
setters/getters to the DatastoreContext. The support for this variable
was added to JMX as well, so that the value can be seen via JConsole.
Moreover, added tests in DatastoreContextTest.

Also fixed a recurring typo in sal-akka-raft. Snapshot was spelled as
snaphot (missing s in shot).

This code was unit tested with different entries in datastore.cfg. Also
tested the case where no special value was provided in datastore.cfg,
and the default value was shown in jconsole.

Change-Id: I024e55b4b47d48037a4e1ceb1006dce32373d75e
Signed-off-by: Shaleen Saxena <ssaxena@brocade.com>