Update patch set 1
Patch Set 1:

This request came from Moiz:

Openflow Plugin has 1 instance of an EntityOwnershipListener - let’s call is OpenflowEntityOwnershipListener (OEOL). Whenever a switch is connected to the controller a RoleChangeListener (RCL) is registered with the OEOL.  When an entity ownership change is received the OEOL calls the RCL which then sets the role of the switch to master/slave. This leads us to the obvious problem that if an ownership change notification was received before the RCL was registered the RCL will not be notified. To workaround this the OEOL now stashes the ownership info so that when a RCL is registered it can notify the RCL of the role.

Solutions,
  1) We notify all listeners whenever a new candidate is registered
  2) We pass a listener to register candidate as before - just to receive the initial notification
  3) Provide a hasOwner and isOwner method on the service which will do a local lookup of the ownership state for an entity

I find option 3 appealing because it is not awkward. What are your thoughts?

Patch-set: 1
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0