Update patch set 1
Patch Set 1:

Apologies for having jumped to conclusions earlier.

Digging deeper into this (and taking Colin's excellent advise), I tried building GBP, NIC, and VTN.

GBP *seems* to just be having issues because it has what *appear* to be unused dependencies on AD-SAL features.  I've pushed a patch here that removes them (and on first blush seems to fix things):

https://git.opendaylight.org/gerrit/#/c/27493/

NIC *appears* to only be having issues indirectly because VTN is having issues.

VTN appears to still depend on the AD-SAL components.  Looking at the VTN release plan:

https://wiki.opendaylight.org/view/OpenDaylight_Virtual_Tenant_Network_(VTN):Beryllium_Release_Plan

This is unsurprising, as they've documented very well that they do not expect to have removed AD-SAL dependencies before M4 offset 2 (2015-11-12).

In order to make sure this gets re-cleaned up, I've filed a patch:

https://git.opendaylight.org/gerrit/#/c/27494/

that reverts this revert, but is marked DO NOT MERGE before 11/13/15, and asks that vtn-dev be consulted before the merge to make sure that things came together as planned on AD-SAL removal.

Patch-set: 1