Update patch set 1
Patch Set 1:

Please look at the following two links,

https://git.opendaylight.org/gerrit/#/c/27419/
https://git.opendaylight.org/gerrit/#/c/26189/14/openflowplugin-impl/src/main/java/org/opendaylight/openflowplugin/impl/role/OpenflowOwnershipListener.java

The first link is to a commit which solves something very similar to what is need for OF clustering, in what I feel is a more efficient manner.

The second link is to an entity ownership listener in the OF lithium branch which currently caches the ownerhip info but could now use the state from EOS.

Also, my understanding is the both Kamal/Kavitha are working on OF clustering one for the Lithium plugin and the other for the Helium one and the code does not seem to be shared. I think atleast some of it can be I think. This way we will avoid requests to solve the same problem in two different ways.

My recommendation is that we abandon this patch.

Patch-set: 1
Reviewer: Gerrit User 795 <795@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0