BUG 4151 : Create a shared actor system
This patch adds an ActorSystemFactory class in clustering commons
which uses the clustered data store configuration and creates a
single shared actor system. The clustered-data store is a better
candidate for this shared system because it does use more configuration
options than the rpc connector which pretty much uses stock configuration.

I also added the ActorSystem as a service on the bundle context, this is
to allow use cases where someone might obtain the actor system from the
bundle context.

Change-Id: I79c14f84c992a2d5ac9c1f1856efbaeba3cc2b77
Signed-off-by: Moiz Raja <moraja@cisco.com>