Update patch set 2
Patch Set 2:

(1 comment)

Both remote-rpc and CDS will shutdown the actor system on module close which will be problematic for on-the-fly module restarts. I think advertising as a service is the right idea but instead of using a static singleton, I think it would be better to create a config module in clustering commons to advertise the ActorSystem service which both remote-rpc and CDS would wire in as a dependency. This way, if we have to restart the ActorSystem, we could simply restart the clustering commons module which would (should) restart the remote-rpc and CDS modules.

Patch-set: 2