Update patch set 2
Patch Set 2:

In comparing the the latest merge job to a previous one, the features-mdsal tests took ~14 min compared to ~2 min previously. This is the cumulative result of each test taking over a minute compared to ~10 sec. However there's no ModuleFactoryBundleTracker timeout errors as was seen before. In fact prior to this patch the ModuleFactoryBundleTracker timeouts caused tests to run for hours (it would timeout 1 min for every ModuleFactory bundle).

So I think this patch did fix the ModuleFactoryBundleTracker timeout issue and the current slower times is not related to the previous ModuleFactoryBundleTracker patch that was merged yesterday. I suspect the slower times are due to the new yang parser. The weather report did mention that it is slower and would cause longer karaf startup times.

Patch-set: 2
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0