Update patch set 1
Patch Set 1:

> Hi Thanh, Rao,
 > it is possible for us to have repository for such modified
 > upstreams in the future?
 > 
 > There is same issue opened for leveldb for ARM architectures?
 > 
 > https://bugs.opendaylight.org/show_bug.cgi?id=3973
 > 
 > It would be nice to have local repo in future and propose to
 > projects
 > to upstrem all changes done in ODL.

Yes, we can upload it to our local thirdparty repo [1]. To get artifacts there though you will need to provide instructions to helpdesk on how to build it OR provide helpdesk with both the jar and source.jar.

Ideally we should get the features into upstream so that we don't have to maintain custom code so I see this as a last resort if the upstream project is unresponsive.

Other thing to keep in mind is how do we version the custom code. I think postfixing -odl is the right thing to do so that we don't confuse the version with upstream but what if release a 2nd or 3rd custom version? maybe postfix -odl-1, -odl-2, etc...?

[1] https://nexus.opendaylight.org/content/repositories/thirdparty/org/fusesource/leveldbjni/

Patch-set: 1