Update patch set 2
Patch Set 2:

It will change the 05-clustering.xml file but there will be migration issues with 01-mdsal I'm sure anyway so it won't really matter :(  When upgrading, we can delete 05-clustering.xml so it's laid down again or maybe we just set the "override" flag to true in the feature. We have the cfg file that's preserved. 

I've been thinking a lot on how to handle/fix the config system upgrade issues in general...

Patch-set: 2
Label: Verified=0