Update patch set 2
Patch Set 2:

Hi Tom,

My inline replies are getting only saved as draft messages. Hence posting my replies here.

Patch Set 1:
(10 comments)
opendaylight/md-sal/sal-distributed-datastore/src/main/java/org/opendaylight/controller/cluster/datastore/ShardManager.java
Line 208:
I don't think we should feedback to the mbean like this - doesn't seem clean. As I commented in the mbean class, the raft actor code needs to check if one is in progress anyway. Also updateShardReplica will be async so setting to false here would be incorrect.
Kalai> The mbean serialization communication is now removed. Serialization is done only in the RaftActor code and not here.

Line 769:
Is the purpose to check if a local shard already exists? If so, then you can check the localShard map.
checkstyle: use '!' instead of "== false" for booleans
kalai> The purpose is to verify the availability of the shard in any of the cluster nodes. i.e. verifying if the shard is listed in the modules.conf.
Updated the verification check for the checkstyle comment.

opendaylight/md-sal/sal-distributed-datastore/src/main/java/org/opendaylight/controller/cluster/datastore/jmx/mbeans/shardmanager/ShardManagerInfo.java
Line 40:
This should be volatile
kalai> Removed the variable added for serialization. Now serialization will be done only in RaftActor.

Line 116:
This check should be done in the ShardManager.
kalai>Done

Line 111:
It would be useful to return the message to the client. I'm not sure how best to do that via JMX, i.e. return a String or throw a runtime ex.
However do we need to do this check here? This has to be checked anyway in the raft code so might as well let that handle it.
kalai> Removed the checks here

Line 120:
This should use the ask pattern to wait for the response.
kalai> Generally the UI request will not be placed in hold for a visible duration of time. Hence for the JMX API requests, the API only posts the request to the ShardManager and does not wait for response for the success/failure of the request.
If an user wants to know the output of the placed request, he can use the JMX API for receiving the localShards available in the node to determine if his request has succeeded/failed.
Or is it mandatory for the JMX to wait for the reply from ShardManager. 


opendaylight/md-sal/sal-distributed-datastore/src/main/java/org/opendaylight/controller/cluster/datastore/messages/UpdateShardReplica.java
Line 12:
Please add class javadocs
kalai> Done

Line 18:
I would prefer separate message classes for add and remove to be consistent as we haven't used this pattern elsewhere.
kalai> Done

Line 41:
Minor thing but Moiz and I have moved to using String concatenation in toString. Please update your IDE (looks like eclipse).
kalai> Done

opendaylight/md-sal/sal-distributed-datastore/src/test/java/org/opendaylight/controller/cluster/datastore/utils/MockConfiguration.java
Line 84:
Minor thing but please add blank line.
kalai> Done

Patch-set: 2