Update patch set 3
Patch Set 3:

Yes - the entire CDS modules would be restarted/recreated which will restart all dependent modules as well (doing it via the config system). This essentially restarts the controller. In your patch to restart when quarantined you would tell the config system to restart the commons module. I think that can be done but if not then just tell the OSGi framework to restart bundle 0.

Patch-set: 3