Update patch set 2
Patch Set 2:

I think there's an easier way to do this rather than adding another concept to RaftActor and that is by using the RaftPolicy. You can build a new DatastoreContext with the custom raft policy class set to one that disables elections. This will keep the shard in the follower state. When you get the success reply you send the normal DatastoreContext to the Shard so it uses the normal RaftPolicy. I'm not sure if a RaftPolicy change is dynamic though in the RaftActor - you may have to do some work there.

Patch-set: 2