Update patch set 2
Patch Set 2:

Cool - thanks for all the testing. I'm looking into lazily creating the SchemaContext in both the ModuleInfoBundleTracker and GlobalBundleScanningSchemaServiceImpl so as not to block bundle activation while parsing the yang models. The important thing is to do the bundle scraping on activation so they yang models are available to build the SchemaContext when it's first needed. This will be done in separate patches.

In the long run it would be ideal to unify ModuleInfoBundleTracker and GlobalBundleScanningSchemaServiceImpl.

Patch-set: 2
Label: Verified=0