Update patch set 7
Patch Set 7:

Path set 7 adds an ActorSystemProviderListener interface for clients to be notified when a new ActorSystem is created. I was originally thinking we could handle ActorSystem restarts via the config system but this has the potential downside of also restarting all apps that use the data store. This adds risk as we'd be relying on all core/app modules gracefully handling close and cleaning up properly. This is something that app devs may not test. 

What do folks think? If it's too troublesome for the data store and RPC to restart via listener callback we can go with the config system restart and take the risks.

Patch-set: 7