Update patch set 3
Patch Set 3:

Tony saw this error when running the SFC distro:

Caused by: java.lang.IllegalStateException: Binding context provider was not set yet
    at com.google.common.base.Preconditions.checkState(Preconditions.java:173)
    at org.opendaylight.controller.config.facade.xml.osgi.YangStoreSnapshot.fromYang(YangStoreSnapshot.java:169)
    at org.opendaylight.controller.config.facade.xml.mapping.attributes.resolving.EnumAttributeResolvingStrategy.parseAttribute(EnumAttributeResolvingStrategy.java:49)
    at org.opendaylight.controller.config.facade.xml.mapping.attributes.fromxml.AttributeConfigElement.resolveValue(AttributeConfigElement.java:46)

I reproduced it. The YangStoreActivator (contained in the config-manager-facade-xml bundle) has a ServiceTracker for the SchemaContextProvider service which is registered by the RefreshingSCPModuleInfoRegistry (the ModuleInfoBackedContext). The SchemaContextProvider  service reference has a couple properties that are set/updated via ModuleInfoBundleTracker -> RefreshingSCPModuleInfoRegistry#updateService - one is the BindingRuntimeContext.  YangStoreActivator ServiceTracker's modifiedService(...) extracts the BindingRuntimeContext property and refreshes its YangStoreService. 

The config-manager bundle is activated before the config-manager-facade-xml bundle, thus the ModuleInfoBundleTracker is created before the YangStoreActivator ServiceTracker. With this patch, bundle activation is complete, the ModuleInfoBundleTracker has already completed the bundle processing and updated the RefreshingSCPModuleInfoRegistry. Therefore the SchemaContextProvider service reference already contains the BindingRuntimeContext. The problem is that the YangStoreActivator ServiceTracker only refreshes the YangStoreService when modifiedService(...) is called which doesn't happen b/c the service is already up-to-date. 

The solution is to also refresh the YangStoreService when it is first created in the addingService(...) method. This fixed the exception in SFC. 

Prior to this patch, the ModuleInfoBundleTracker scraping was done well after the YangStoreActivator ServiceTracker was created so it always received at least 1 modifiedService(...) call.

Patch-set: 3
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0