Update patch set 7
Patch Set 7:

(14 comments)

Please add unit tests for ShardManager which processes CreateShardReplica/RemoveShardReplica. Also the antonym for Remove would be Add so the names should be add/remove. If you use Create the antonym would be Destroy and I don't think that naming fits what we're trying to do. It would be nice to keep naming consistent.

Patch-set: 7