Update patch set 1
Patch Set 1:

(1 comment)

Alright. I'm not thinking of this as an API. The reason is that we would be leaking too much implementation detail if we did so. At this time apps will not contribute this. My thoughts are as follows, we implement some simple strategies in the datastore. Then using a configuration file we associate entity-types with strategies. What do you think?

Patch-set: 1
Label: Verified=0