Update patch set 2
Patch Set 2:

Yeah - I wasn't thinking API either. Static config is one approach. We could go with a shared file like module-shards but then app devs have to modify it which isn't clean. How about scraping contrib properties files from bundles? The file would specify the entity type, strategy class and params. I prefer the bean pattern with setters - then we can use the BeanIntrospector to do the reflection work.

However that would be a bit limiting if we allow for custom strategy classes where it might need dependencies that we can't provide. In that case you want the app to create the instance and it could be delivered via OSGi service registration. The downside is that we wouldn't have the strategy on startup.

I think the properties file approach is the way to go. I don't think there will be a need for custom strategies. The 2 strategies I can see are FCFS and load balanced.

Patch-set: 2