BUG-2187: Add Server - Leader Implementation
Processes addServer request from the follower, forwards the request
to the shard leader, if not the leader.

The follower shard replica data is brought to sync with leader by
installing the snapshot from the shard leader.
On sucessful application of snapshot data, this voting but not initialized member
is transitioned to voting member.
New server configuration is persisted and replicated to majority of
the followers and responds back with OK message to the shard follower.

In case where the leader is unable to sync data to the follower in a
configured time period, TIMEOUT message is responded back to the shard follower
without adding/persisting the new server configuration.

Change-Id: I9a3870d14bb6ad532ff64f315b2e2000d8b803e2
Signed-off-by: Rajesh_Sindagi <Rajesh_Sindagi@dell.com>