Update patch set 6
Patch Set 6:

Can you also change the default auto-down-after-unreachable-after setting? I was thinking  like 30s. This should eliminate the intermittent car test failures.

Relying on exception message text is obviously fragile so it would be nice to come up with a unit test for the QuanrantinedMonitorActor so we can detect if it changes when we upgrade akka. This would likely be tricky - I think you'd have to some how stop traffic between 2 actor systems. This can be investigated and done, if feasible, in a separate patch - no need to hold this one up.

Patch-set: 6
Label: Verified=0