Clean up duplicate/unused dependencies and properties
Remove dependencies and properties provided in odlparent (with the
same versions).

org.json.version in features/mdsal/pom.xml is unused.

corsfilter.version is only used in controller so replace it with the
version.

(All this will allow a number of properties to be removed from
odlparent.)

Some of the remaining properties which appear unused in controller are
used in reservation.

Change-Id: I07e9f2298ebd008d82b22b156dc2ddce50151641
Signed-off-by: Stephen Kitt <skitt@redhat.com>