Update patch set 1
Patch Set 1:

(1 comment)

What is the reason for persisting the initializing/initialized state? 

The way I thought we would do it is to persist the new list of shard names when AddServer successfully completes. Also persist it as a snapshot not as a log entry. On recovery, we could add the shard replicas to the Configuration and then create the local shards on RecoveryComplete instead of in the constructor. I think this would simply the code. The actor isn't ready to receive normal messages until after recovery completes.

Patch-set: 1